# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vehicle Management System Api Documentation",
        "description": "L5 Swagger API Documentation for Laravel Project Pharmacy",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/boats": {
            "get": {
                "tags": [
                    "Boats"
                ],
                "summary": "Get List of all Boats",
                "description": "Returns List of all boats",
                "operationId": "boatIndex",
                "responses": {
                    "200": {
                        "description": "A list of boats"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page Not Found"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Boats"
                ],
                "summary": "Store a boat",
                "description": "Store boat",
                "operationId": "boatStore",
                "requestBody": {
                    "description": "Pass boat data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "make",
                                    "model",
                                    "year",
                                    "seats",
                                    "color",
                                    "vin",
                                    "current_mileage",
                                    "service_interval",
                                    "next_service"
                                ],
                                "properties": {
                                    "make": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "KIA"
                                    },
                                    "model": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Sportage"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "2023"
                                    },
                                    "seats": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "4"
                                    },
                                    "color": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "White"
                                    },
                                    "vin": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "KUCHBELIKH465452"
                                    },
                                    "current_mileage": {
                                        "type": "decimal",
                                        "format": "text",
                                        "example": "11.5"
                                    },
                                    "service_interval": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "3000KM"
                                    },
                                    "next_service": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "6000KM"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Display a boat"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page Not Found "
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/boats/{id}": {
            "get": {
                "tags": [
                    "Boats"
                ],
                "summary": "Get a Boat",
                "description": "Display a boat",
                "operationId": "boatShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the boat to show",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "show a boat"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page Not Found"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Boats"
                ],
                "summary": "Update a boat",
                "description": "Update boat",
                "operationId": "boatUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the boat to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass boat data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "make",
                                    "model",
                                    "year",
                                    "seats",
                                    "color",
                                    "vin",
                                    "current_mileage",
                                    "service_interval",
                                    "next_service"
                                ],
                                "properties": {
                                    "make": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "KIA"
                                    },
                                    "model": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Sportage"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "2023"
                                    },
                                    "seats": {
                                        "type": "integer",
                                        "format": "text",
                                        "example": "4"
                                    },
                                    "color": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "White"
                                    },
                                    "vin": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "KUCHBELIKH465452"
                                    },
                                    "current_mileage": {
                                        "type": "decimal",
                                        "format": "text",
                                        "example": "11.5"
                                    },
                                    "service_interval": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "3000KM"
                                    },
                                    "next_service": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "6000KM"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update a boat"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page Not Found "
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Boats"
                ],
                "summary": "Get a Boat",
                "description": "Delete a boat",
                "operationId": "boatDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the boat to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete a boat"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page Not Found"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}